Initial check-in.
  * This is a bit of a grab-bag of things intended to grow towards
being an implementation of the parts of the CLIM standard that
actually make sense to me.

  * Anything in the root directory is either a test driver program
or support code for the test driver, more to have something that
works to start from than to become part of the final system.

  * Anything in tests/ is stale test code for something that has
been swiped from another project (or earlier incarnation of
nq-clim).

  * There is an implementaion of SPACE-REQUIREMENTs in layout/,
not because we are anywhere near implementing the layout protocol
but because it was a convenient abstraction when setting up some
of the support code in its original context.

  * And the rest is a "spike" for drawing mediums, the most
fundamental output protocol in CLIM.  It's thin, but it works.