port/port-discovery: How to find (and possibly make) a suitable port.
  * There is some clever magic here, along with an nq-clim private
extension interface (not part of the specification) for creating
new port instances.  For the most part, however, this is...

  * Actually, it's two different things, which is not right.  This
file both manages the *ALL-PORTS* list for MAP-OVER-PORTS, it also
covers FIND-PORT and *DEFAULT-SERVER-PATH*.  In order for this to
work, CREATE-PORT would need an :AROUND method to update
*ALL-PORTS*, removing that responsibility from FIND-PORT.

  * FIND-PORT requires MAP-OVER-PORTS.  MAP-OVER-PORTS uses
*ALL-PORTS*, and maintaining *ALL-PORTS* would require methods on
both DESTROY-PORT and CREATE-PORT, necessitating three files in
place of one.  We might go there at some point, but not now.