clx-interface: Restructure window property configuration.
  * Move setting the window background color and event mask to
separate SETF forms, rather than passing them as parameters to
XLIB:CREATE-WINDOW, as preparation for using a mirrored sheet,
which won't allow us to pass such parameters to start with (or,
possibly, ever).

  * Lose the setting for the window border color.  The window
seems to either have a zero-width border by default or the window
manager sets it down to that during reparenting, so we don't need
to set it.