backend/clx/graft: Fix CLX-GRAFT superclass order.
  * SHEET, being the most-abstract class, should be LAST on the
list, as a number of GFs have methods defined to do something
default on SHEET and expect to be overridden by subclasses or
mixins.

  * In retrospect, requiring a specific ordering for mixins vs. a
base class is weak design, but I don't currently have a good angle
for sorting it out and it takes a back seat to getting the basic
system functional.