clx-interface: Create and use a CLX-GRAFT to find the root window.
  * One step closer to operating in terms of CLIM rather than CLX,
we create a CLX-GRAFT (through the private creation interface
rather than FIND-GRAFT or whatever) and request its direct-mirror
as the root window.

  * The advantage here is that as soon as we have the ability to
REALIZE-MIRROR for CLX-PORT on a normal MIRRORED-SHEET-MIXIN, we
can use that to create the window instead of using
XLIB:CREATE-WINDOW directly.  Add being able to get a MEDIUM for a
mirrored sheet and some event handling and we've got enough of the
"silica" layer complete to start in on the higher-level bits.