rendering_1: Extract HANDLE-ONE-EVENT from RUN-EVENT-LOOP.
  * Now that we have an explicit loop and loop exit control we can
extract the logic to handle a single event to its own function,
away from the logic that runs the full event loop.

  * Note that, while this leaves us with a RUN-EVENT-LOOP with a
single responsibility, the HANDLE-ONE-EVENT function is still
badly overloaded.  And is about to become the inflection point for
moving towards reified event objects (CLIM-style or not, these can
only help at this point).