frame/standard-application-frame: Initial rough-in for the base frame class.
  * This is several things, and not much at all.  It establishes
the common superclass for most application frames.  It provides a
hook onto which we can hang some functionality.  And it consists
of a single slot with an accessor and an initarg.  It is the thin
end of the wedge for getting the frame, frame-manager, panes, top
level loop in place (basically the non-stream-based, non-gadget
parts of the application layer).