geometry/standard-rectangle-set: Starting in on composite regions.
  * I had been hoping to completely avoid implementing composite
regions, but over the past couple of days it has become obvious
that the CLIM architecture is heavily biased towards maintaining
its own windowing hierarchy that only touches a host system where
absolutely necessary, and therefore needs to keep track of sheet
regions including visible sheet regions, which means composite
regions.

  * Every previous time that I've tried to implement composite
regions, I've failed, for various excuses.  This time, I've found
enough of a hint as to how to implement them correctly, in the
form of five pages out of an 1174-page book, that I'm confident
that I will end up with a usable implementation.

  * This commit provides the implementation of the innermost part
of the UNION operation for sets of rectangular regions.  With
test cases, no less.