tests/standard-rectangle-set-tests.lisp: Two more tests for UNITE-X-SPAN-SETS.
  * We don't (AFAIK) currently call UNITE-X-SPAN-SETS with an
empty first or second set, since the semantics are to return the
non-empty set at that point so we just skip the call, but the
INTERSECTION and DIFFERENCE operations have differing semantics
for one or the other of the sets being empty, and we're looking to
use common toplevel logic for all three operations.

  * Test to make sure that UNITE-X-SPAN-SETS does something
reasonable when either of the input sets are empty.