geometry/standard-rectangle-set: Factor out more generic X-span handling.
  * Did I say that things were going to become more interesting?
Now's the time.

  * Rename INTERSECT-X-SPAN-SETS to OPERATE-ON-X-SPAN-SETS, and
add an extra parameter for the operation.  Whenever the earlier
version of the function would have collected an X-span (either for
real or in commented-out code), do so, but only if calling the
operation (with flags indicating if the span is a member of each
set) returns true.

  * Add a new definition of INTERSECT-X-SPAN-SETS defined in terms
of OPERATE-ON-X-SPAN-SETS and a new INTERSECTION-OPERATION, which
is pretty much the definition of intersection.