geometry/standard-rectangle-set: The start of composite region differences.
  * Start by adding a series of test cases for calculating X-span
differences.  This is the set-theoretic notion of "difference",
meaning that the result is those parts of the first set that are
not parts of the second set.  Then define the required operation
function to make the tests pass (DIFFERENCE-OPERATION).

  * One of the things that I like about the way this whole mess
wound up is that the basic operations are really simple and relate
easily to the set-theoretic definitions.