geometry/nowhere: Initial rough-in, the empty region.
  * This is the region comprising of absolutely no points.  It is
unbounded, and has no dimensions.  The main thing to note is that
it has a privileged "singleton" instance which is a constant (as
per CLIM II 3.1).

  * Here we establish the class, the constant instance, print-read
consistency, and externalizability of literal uses in compiled
files.

  * We do not yet add it to the build order, as I have yet to
fully sort out what its dependencies are and what it depends on,
especially as we move towards the region predicate and region
composition protocols.