geometry/nowhere: Define TRANSFORM-REGION methods for +NOWHERE+.
  * TRANSFORM-REGION typically needs to specialize on both the
transformation type and the region type, with a couple of notable
exceptions.  The first is the identity transformation.  The second
is the empty region (also the full region, +EVERYWHERE+, but I
have yet to find a need to implement it).

  * Implement TRANSFORM-REGION (and UNTRANSFORM-REGION) methods
for +NOWHERE+, which return +NOWHERE+ because there are no points
to transform.