geometry/everywhere: It turns out that we need +EVERYWHERE+ after all
  * It turns out that mediums may have a clipping region, and that
the default clipping region is +EVERYWHERE+.  In the medium term,
this means a small pile of additional complexity in the geometry
layer.  In the short term, this at least means a copy/paste
routine on the implementation of +NOWHERE+ to produce +EVERYWHERE+.

  * We could refactor so that we have an UNBOUNDED-REGION class
that holds the common MAKE-LOAD-FORM, PRINT-OBJECT, and transform
logic, but we'd still need the region-region operations to
specialize on a suitably distinct class.  Possibly a factoring for
the future.