clim-sys/named-constant-mixin: A mixin for the common "constant class" idiom
  * Already in NQ-CLIM we have three cases where a singleton
instance of some class is defined to be a constant.  The actual
implementation of these cases is somewhat verbose...  and in one
case, inconsistent.  And we have at least eleven more in the
offing.

  * Define a mixin class and a macro to paper over almost all of
the work involved in this idiom.  The idea is to use the mixin as
part of the singleton (or flyweight) class, then use the macro to
define the actual constants.  The class definition still needs to
be made available at compile-time, and any documentation of the
constant needs to be applied afterwards, but a lot of the
boilerplate just goes away.