geometry/identity-transformation: Convert to NAMED-CONSTANT-MIXIN
  * This was the inconsistent implementation of the idiom, so its
behavior changed slightly.  Notably, the original did not have a
PRINT-OBJECT method, so it rather lost print/read consistency.
The new version does have such a method.

  * Another difference to the previous two conversions is that
there is no documentation supplied for the named constant.  It did
not disappear during the conversion, it was never supplied in the
first place.