geometry/everywhere: Implement region composition with +EVERYWHERE+
  * Slightly more complex than the equivalent code for +NOWHERE+,
and it turns out that one of the REGION-DIFFERENCE cases is not
implementable per the CLIM II spec because REGION-DIFFERENCE must
return either a "simple region" or a REGION-SET, and REGION-SET is
a bounded region, but the negative region produced by taking the
difference of some other region from +EVERYWHERE+ returns an
unbounded region with a hole in it.

  * If we ever trigger the error, we'll consider doing something
about it, but until then signalling an error is sufficient.