sheet/mirrored-sheet-mixin: Realize the mirror when the sheet is grafted
  * A few crazy things here.  First, we declare that grafting a
sheet is sufficient to cause the mirror to be realized.  Second,
we use a :BEFORE method on NOTE-SHEET-GRAFTED, when the extension
protocol is to use an :AFTER method, in the hopes that any
interaction with the PERMANENT-MEDIUM-SHEET-OUTPUT-MIXIN will come
out favorably this way.  Third, since the mirror is already
realized, clx-inteface needs to not realize it again, just to ask
what the mirror currently is.