backend/clx/medium: Initial implementation of medum {,de}allocation
  * Deallocation is so straightforward as to almost be a one-liner.

  * Allocation is a bit of a nightmare, complicated by having to
match what is currently done in the test program because of not
having suitable control over medium properties yet, and then
limitations in CLX to do with getting to a "screen" structure from
a window, followed by what could be argued to be an actual *bug*
in CLX (screen root windows aren't interned?  what?).

  * Most, possibly even all of the damage will go away once we
have the ability to set color inks and line-styles.