medium/graphics-method: Initial rough-in for basic drawing functions
  * The basics for DRAW-POINT*, DRAW-LINE*, and friends.

  * Two issues here.  First, the spec breaks down when it comes to
rectangles and polygons, with respect to the CLOSED and FILLED
keyword arguments.  Second, there is a clear need for these
functions (and others) to accept keyword arguments to pass to the
underlying MEDIUM- functions.

  * There is also an opportunity for excellence, in limiting the
drawing options accepted per graphics-method to those that make
sense for the graphics-method, and setting up the arglist hinting
appropriately.

  * But there's definitely a mess here to clean up before we get
as far as actually rendering text.