backend/clx/medium: Cache the window colormap when allocating a medium
  * Two reasons for this.  One is that it saves us a step.  The
other is that PIXMAPs are the other drawables that we're going to
need to support, and they don't HAVE colormaps, but are defined in
terms of another medium, which will (because it's a window medium
or because it's a pixmap medium that was defined in terms of a
medium which has a colormap).

  * So, yeah, cache the colormap.  We'll be using it soon to look
up colors, so that we can use them for actual non-black output.
How cool is that?