backend/clx/medium: Simplify medium allocation
  * Now that we are updating the GCONTEXT based on the INK and
BACKGROUND components of the MEDIUM, and the initial cached values
are not part of the gamut for the actual components, we no longer
need to go out of our way to find the appropriate SCREEN in order
to find the black and white pixel values that we were using for
the GCONTEXT foreground and background.  We can leave them at
their defaults and trust the cache invalidation mechanism to set
them up as required.