medium/graphics-method: Some keyword-arglist-hacking utilities
  * This is plausibly not the right place for these, and if we
ever find down the line that we need this functionality in another
package then they get kicked straight over to CLIM-SYS.

  * One item here is a keyword parameter specifier parser, which
returns the various appropriately-defaulted bits of a keyword
parameter as values.

  * The other uses that information to process a set of keyword
parameter specifiers into a new set of keyword parameter
specifiers and some code to construct an "equivalent" parameter
list, suitable for using with DESTRUCTURING-BIND or passing to
APPLY.

  * Why all this mess?  Because we need to be able to split out
two different groups of keyword arguments from a function, each to
be used in a separate function call, and if we do it this way
instead of using :ALLOW-OTHER-KEYS we get error checking and
working arglist hinting.