medium/graphics-method: Represent drawing-option arguments explicitly
  * Now that all of the preliminaries are out of the way, use our
new keyarg-mangling logic to explicitly bind ONLY the options that
are valid for a given graphics-method, and to pass a minmal set of
them (in the case of duplicated arguments) to
INVOKE-WITH-DRAWING-OPTIONS.

  * What are the upshots?  First off, arglist hints!  Second, we
can now turn around and do the same thing for keyword parameters
to the actual MEDIUM-DRAW-whatever functions.

  * Oh, and while we're here, since we're now using them, make the
KEY-ARGS and ALLOWED-OPTIONS macro parameters required.