clx-interface: If a frame has FRAME-PANES set, adopt it to the toplevel sheet
  * Lousy way to express it, but I'm pretty close to wiped out
right now.

  * When setting up the toplevel sheet, getting it mapped to the X
server, and all that, basically when pretending to be a frame
manager without the actual frame-manager object, if the frame has
FRAME-PANES set, have the toplevel sheet adopt the pane.

  * This is the hook that should allow us to move towards "real"
pane allocation and layout, better management of mediums, and once
we get event handling under control, repaint.  But it's still an
ugly hack.