rendering_1: Handle events in terms of FRAME, not *DISPLAY*
  * Okay, in a couple of ways this looks like a bit of a hack, but
largely because the main system is missing proper frame management
and any conception of event handling.

  * Pass the FRAME down through the toplevel event-loop function
to the two places where we used *DISPLAY* for something.

  * Where we used *DISPLAY*, obtain the PORT (going via the
FRAME-PANES slot because it's not yet supported on application
frames), and then use NQ-CLIM/BACKEND/CLX/PORT:CLX-PORT-DISPLAY to
obtain the actual display.  The upside is that the places where we
need to do this whole dance are event-handling functions, and that
means that they're going to get cleaned up at some point anyway.