backend/clx/medium: Specialize PORT and SHEET in methods dealing with mediums
  * Tighten things up a bit here.  It makes no sense to work with
a CLX-MEDIUM with a port that is not a CLX-PORT.  It makes no
sense to allocate a medium that is compatible with a sheet that
can't have a medium associated with it.  It makes no sense to
associate a medium to or disassociate a medium from a sheet that
can't have a medium associated with it.