sheet/standard-sheet-input-mixin: Initial implementation
  * There are four major parts here.

  * First is an implementation of SHEET-EVENT-QUEUE, which
presumes that the parent sheet has an implementation of
SHEET-EVENT-QUEUE that can return a usable result.  There's also a
mention of a couple of other possible implementation strategies,
both for comparison and in case we decide that the current setup
is insufficient to requirements.

  * Second is an implementation of DISPATCH-EVENT that simply
queues the event.  As noted in the comments, this is only really
correct for DEVICE-EVENTs, and various other event types may need
special handling, but we don't yet have any actual event classes
yet (we're using simple plists as events) so always queueing is
appropriate.

  * Third is a dummy implementation of HANDLE-EVENT which simply
ignores the event.  We're probably going to have to fill this out
over time, but again, we don't have event classes yet.

  * Fourth is an implementation of the event queue protocol that
reflects all calls to the SHEET-EVENT-QUEUE of the sheet.  Simple,
to the point, and not really expected to change going forward.