backend/clx/frame-sheet: Tie the frame-sheet into the input protocol
  * Two pieces here.  The first is simply that the frame-sheet
needs to handle input, thus needs to include the standard input
mixin.  The second is that this is the highest non-graft mirror
for the application, and thus the natural place to terminate our
search for the per-frame event queue, so we just put a slot on the
class hold the queue and set up SHEET-EVENT-QUEUE as a reader,
overriding the (less-specific) method on the standard input mixin.