rendering_1: Convert to new event-handling infrastructure
  * Okay, major progress towards "proper" input handling.  Two
things going on here, and a bit of a caveat.

  * First, we add STANDARD-SHEET-INPUT-MIXIN to our MAZE-PANE
class, so that it's capable of participating in the whole event
handling dance.

  * Second, we remove the old event-fetching logic that queried
the X server directly and replace it with a use of
PROCESS-NEXT-EVENT and some of the event-queue functions.

  * Now, our main caveat: we "should" be bouncing the events back
into the handling infrastructure by way of HANDLE-EVENT, but that
requires being able to determine, for each event (and each event
type!), which "client" (typically, sheet) is supposed to handle
the event, and we're not set up for that with our current plist
event mockeries.